refactor: stepsection 수정
- cardlist constanse 경로수정
- stepsection css 수정
- carousel css 수정
