+++ apps/pick-learn/src/entities/guide/model/constants.ts
