+++ apps/pick-learn/src/entities/guide/ui/StepCardList.tsx
@@ -1,9 +1,9 @@
 import { stepList } from '../model/constants';
-import StepCard from '@/widgets/home/ui/StepCard';
+import StepCard from '@/entities/guide/ui/StepCard';
 
 export default function StepCardList() {
     return (
-        <ul className='col-span-1 md:col-span-8 md:col-start-6 order-2 md:order-none  gap-6 hidden md:flex'>
+        <ul className='col-span-1 lg:col-span-7 xl:col-span-8 gap-6 lg:justify-end hidden md:flex '>
             {stepList.map(({ icon: Icon, ...item }) => (
                 <StepCard key={item.id}>
                     <StepCard.Icon>
