+++ apps/pick-learn/src/shared/ui/Wrapper/StepWrapper.tsx
@@ -1,0 +1,17 @@
+import { cn } from '@repo/ui/lib/utils';
+
+export default function StepWrapper({
+    children,
+    className,
+}: {
+    children: React.ReactNode;
+    className?: string;
+}) {
+    return (
+        <section
+            className={cn('w-full py-[5rem] bg-gray-100 md:px-0', className)}
+        >
+            {children}
+        </section>
+    );
+}
