+++ apps/pick-learn/src/shared/ui/index.ts
@@ -1,5 +1,6 @@
 import Label from './Label';
 import Tag from './Tag';
 import MainWrapper from './Wrapper/MainWrapper';
+import StepWrapper from './Wrapper/StepWrapper';
 
-export { Label, Tag, MainWrapper };
+export { Label, Tag, MainWrapper, StepWrapper };
