+++ apps/pick-learn/src/views/home/ui/index.ts
@@ -1,5 +1,5 @@
 import StepSection from './StepSection';
-import StepCardList from './StepCardList';
-import StepCardCarousel from './StepCardCarousel';
+import StepCardList from '../../../entities/guide/ui/StepCardList';
+import StepCardCarousel from '../../../entities/guide/ui/StepCarouseList';
 
 export { StepSection, StepCardList, StepCardCarousel };
